backend/requirements.txt
appnope==0.1.3
asgiref>=3.6.0
astroid==2.15.4
asttokens==2.2.1
backcall==0.2.0
black==22.3.0
click==8.1.2
decorator==5.1.1
dill==0.3.6
Django>=4.1.10
django-extensions==3.2.3
django-cors-headers==4.2.0
django-spa==0.3.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils==0.19
executing==1.2.0
flake8==4.0.1
ipython==8.13.1
isort==5.12.0
jedi==0.18.2
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy==1.2.0
mypy-extensions==1.0.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
psycopg2-binary==2.9.9
pyflakes==2.4.0
Pygments==2.15.1
pylint==2.17.3
pyserial==3.5
Send2Trash==1.8.2
six==1.16.0
sqlparse>=0.4.4
stack-data==0.6.2
thonny==4.0.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
whitenoise==5.3.0
wrapt==1.15.0
